Adding maximum file limit feature
This script is great, except for one limitation - user cannot enter a maximum file size. I have added that functionality. User will input a parameter named maxSize (by default it will be 1024), which will be counted as kilobyte. Then the file size will be compared with user's given size limit. If the file size exceeds the limit, the file is not uploaded and an error message shows.